refactor: Extract schemas + split models into packages (#167)
Phase 2a schema/model restructure:
- Extract 78 inline Pydantic schemas from 8 route files into schemas/ package
- Split monolithic models.py (19 models) into models/ package with 7 domain files
- Add TimestampMixin applied to 13 models
- All 620 tests pass, zero import breakage
