+++ ai_ready_rag/api/documents.py
@@ -2,7 +2,6 @@
 
 import asyncio
 import logging
-from datetime import datetime
 from typing import Annotated
 
 from fastapi import (
@@ -15,13 +14,26 @@ from fastapi import (
     UploadFile,
     status,
 )
-from pydantic import BaseModel
 from sqlalchemy.orm import Session
 
 from ai_ready_rag.config import get_settings
 from ai_ready_rag.core.dependencies import get_current_user, require_admin
 from ai_ready_rag.db.database import SessionLocal, get_db
 from ai_ready_rag.db.models import Document, User
+from ai_ready_rag.schemas.document import (
+    BulkDeleteRequest,
+    BulkDeleteResponse,
+    BulkDeleteResult,
+    BulkReprocessRequest,
+    BulkReprocessResponse,
+    CheckDuplicatesRequest,
+    CheckDuplicatesResponse,
+    DocumentListResponse,
+    DocumentResponse,
+    DuplicateInfo,
+    ReprocessRequest,
+    TagUpdateRequest,
+)
 from ai_ready_rag.services.document_service import DocumentService
 
 logger = logging.getLogger(__name__)
@@ -134,95 +146,6 @@ async def process_document_task(
                 logger.exception("Failed to update document status after error")
         finally:
             db.close()
-
-
-class TagInfo(BaseModel):
-    id: str
-    name: str
-    display_name: str
-
-    class Config:
-        from_attributes = True
-
-
-class TagUpdateRequest(BaseModel):
-    tag_ids: list[str]
-
-
-class ReprocessRequest(BaseModel):
-    enable_ocr: bool = True
-    force_ocr: bool = False
-    ocr_language: str = "eng"
-
-
-class DocumentResponse(BaseModel):
-    id: str
-    original_filename: str
-    filename: str
-    file_type: str
-    file_size: int
-    status: str
-    title: str | None
-    description: str | None
-    chunk_count: int | None
-    page_count: int | None
-    word_count: int | None
-    processing_time_ms: int | None
-    error_message: str | None
-    tags: list[TagInfo]
-    uploaded_by: str
-    uploaded_at: datetime
-    processed_at: datetime | None
-
-    class Config:
-        from_attributes = True
-
-
-class DocumentListResponse(BaseModel):
-    documents: list[DocumentResponse]
-    total: int
-    limit: int
-    offset: int
-
-
-# Duplicate check models
-class CheckDuplicatesRequest(BaseModel):
-    """Request for pre-upload duplicate check."""
-
-    filenames: list[str]
-
-    @property
-    def validated_filenames(self) -> list[str]:
-        """Ensure at least one filename provided."""
-        if not self.filenames:
-            raise ValueError("At least one filename required")
-        return self.filenames
-
-
-class DuplicateInfo(BaseModel):
-    """Information about a duplicate file."""
-
-    filename: str
-    existing_id: str
-    existing_filename: str
-    uploaded_at: datetime
-
-
-class CheckDuplicatesResponse(BaseModel):
-    """Response from duplicate check."""
-
-    duplicates: list[DuplicateInfo]
-    unique: list[str]
-
-
-class DuplicateErrorDetail(BaseModel):
-    """Structured 409 error response."""
-
-    detail: str
-    error_code: str = "DUPLICATE_FILE"
-    existing_id: str
-    existing_filename: str
-    uploaded_at: datetime
 
 
 @router.post("/check-duplicates", response_model=CheckDuplicatesResponse)
@@ -457,28 +380,6 @@ async def delete_document(
     return None
 
 
-class BulkDeleteRequest(BaseModel):
-    """Request body for bulk delete."""
-
-    document_ids: list[str]
-
-
-class BulkDeleteResult(BaseModel):
-    """Result for a single document deletion."""
-
-    id: str
-    status: str  # "deleted" or "failed"
-    error: str | None = None
-
-
-class BulkDeleteResponse(BaseModel):
-    """Response for bulk delete operation."""
-
-    results: list[BulkDeleteResult]
-    deleted_count: int
-    failed_count: int
-
-
 @router.post("/bulk-delete", response_model=BulkDeleteResponse)
 async def bulk_delete_documents(
     request: BulkDeleteRequest,
@@ -657,20 +558,6 @@ async def reprocess_document(
     )
 
     return document
-
-
-class BulkReprocessRequest(BaseModel):
-    """Request body for bulk reprocess."""
-
-    document_ids: list[str]
-
-
-class BulkReprocessResponse(BaseModel):
-    """Response for bulk reprocess operation."""
-
-    queued: int
-    skipped: int
-    skipped_ids: list[str]
 
 
 @router.post(
