+++ ai_ready_rag/db/models/__init__.py
@@ -1,0 +1,52 @@
+"""SQLAlchemy ORM models.
+
+Re-exports all models for backward compatibility.
+Import from here: ``from ai_ready_rag.db.models import User, Tag, Document``
+"""
+
+from ai_ready_rag.db.models.admin import (
+    AdminSetting,
+    AuditLog,
+    ReindexJob,
+    SettingsAudit,
+    SystemSetup,
+)
+from ai_ready_rag.db.models.base import TimestampMixin, document_tags, generate_uuid, user_tags
+from ai_ready_rag.db.models.cache import (
+    CacheAccessLog,
+    EmbeddingCache,
+    ResponseCache,
+    WarmingFailedQuery,
+    WarmingQueue,
+    WarmingSSEEvent,
+)
+from ai_ready_rag.db.models.chat import ChatMessage, ChatSession
+from ai_ready_rag.db.models.document import Document
+from ai_ready_rag.db.models.rag import CuratedQA, CuratedQAKeyword, QuerySynonym
+from ai_ready_rag.db.models.user import Tag, User
+
+__all__ = [
+    "TimestampMixin",
+    "generate_uuid",
+    "user_tags",
+    "document_tags",
+    "User",
+    "Tag",
+    "Document",
+    "ChatSession",
+    "ChatMessage",
+    "AuditLog",
+    "AdminSetting",
+    "SystemSetup",
+    "SettingsAudit",
+    "ReindexJob",
+    "ResponseCache",
+    "EmbeddingCache",
+    "CacheAccessLog",
+    "WarmingQueue",
+    "WarmingFailedQuery",
+    "WarmingSSEEvent",
+    "QuerySynonym",
+    "CuratedQA",
+    "CuratedQAKeyword",
+]
