+++ ai_ready_rag/schemas/__init__.py
@@ -1,0 +1,255 @@
+"""Pydantic schemas for API request/response models.
+
+Import individual modules for schemas with conflicting names:
+- chat.BulkDeleteRequest / chat.BulkDeleteResponse (session_ids)
+- document.BulkDeleteRequest / document.BulkDeleteResponse (document_ids)
+- admin.BulkDeleteRequest (job_ids)
+"""
+
+from ai_ready_rag.schemas.admin import (
+    AdvancedSettingsRequest,
+    AdvancedSettingsResponse,
+    ArchitectureInfoResponse,
+    CacheClearResponse,
+    CacheSeedRequest,
+    CacheSeedResponse,
+    CacheSettingsRequest,
+    CacheSettingsResponse,
+    CacheStatsResponse,
+    CacheWarmRequest,
+    CacheWarmResponse,
+    ChangeEmbeddingRequest,
+    ChangeEmbeddingResponse,
+    ChangeModelRequest,
+    ChangeModelResponse,
+    ChatModelInfo,
+    ClearKnowledgeBaseRequest,
+    ClearKnowledgeBaseResponse,
+    ComponentHealth,
+    CuratedQACreate,
+    CuratedQAListResponse,
+    CuratedQAResponse,
+    CuratedQAUpdate,
+    DetailedHealthResponse,
+    DocumentParsingInfo,
+    EasyOCRStatus,
+    EmbeddingsInfo,
+    FileStats,
+    InfrastructureStatus,
+    KnowledgeBaseStatsResponse,
+    KnowledgeBaseSummary,
+    LLMSettingsRequest,
+    LLMSettingsResponse,
+    ManualWarmingRequest,
+    ModelInfo,
+    ModelLimits,
+    ModelLimitsResponse,
+    ModelsResponse,
+    OCRStatus,
+    ProcessingOptionsRequest,
+    ProcessingOptionsResponse,
+    ProcessingQueueStatus,
+    RAGPipelineStatus,
+    RecoverResponse,
+    ReindexEstimate,
+    ReindexFailureInfo,
+    ReindexFailuresResponse,
+    ReindexJobResponse,
+    ResumeReindexRequest,
+    RetrievalSettingsRequest,
+    RetrievalSettingsResponse,
+    SettingsAuditEntry,
+    SettingsAuditResponse,
+    StartReindexRequest,
+    SynonymCreate,
+    SynonymListResponse,
+    SynonymResponse,
+    SynonymUpdate,
+    TesseractStatus,
+    TopQueryItem,
+    TopQueryResponse,
+    WarmFileResponse,
+    WarmingJobListResponse,
+    WarmingJobResponse,
+    WarmingQueueJobResponse,
+    WarmingQueueListResponse,
+    WarmRetryRequest,
+)
+from ai_ready_rag.schemas.auth import (
+    LoginRequest,
+    LoginResponse,
+    SetupRequest,
+    UserBasicResponse,
+)
+from ai_ready_rag.schemas.chat import (
+    ConfidenceInfo,
+    DeleteSessionResponse,
+    MessageCreate,
+    MessageListResponse,
+    MessageResponse,
+    SendMessageResponse,
+    SessionCreate,
+    SessionListItem,
+    SessionListResponse,
+    SessionResponse,
+    SessionUpdate,
+    SourceInfo,
+)
+from ai_ready_rag.schemas.document import (
+    BulkReprocessRequest,
+    BulkReprocessResponse,
+    CheckDuplicatesRequest,
+    CheckDuplicatesResponse,
+    DocumentListResponse,
+    DocumentResponse,
+    DuplicateErrorDetail,
+    DuplicateInfo,
+    ReprocessRequest,
+    TagInfo,
+    TagUpdateRequest,
+)
+from ai_ready_rag.schemas.experimental import (
+    GeneratePresentationRequest,
+    PresentationResponse,
+    SlideResponse,
+)
+from ai_ready_rag.schemas.setup import (
+    CompleteSetupRequest,
+    CompleteSetupResponse,
+    SetupStatusResponse,
+)
+from ai_ready_rag.schemas.tag import (
+    TagCreate,
+    TagResponse,
+    TagUpdate,
+)
+from ai_ready_rag.schemas.user import (
+    TagAssignment,
+    UserCreate,
+    UserResponse,
+    UserUpdate,
+)
+
+# Note: BulkDeleteRequest and BulkDeleteResponse exist in chat, document, and admin
+# modules with different fields. Import them directly from their respective modules:
+#   from ai_ready_rag.schemas.chat import BulkDeleteRequest, BulkDeleteResponse
+#   from ai_ready_rag.schemas.document import BulkDeleteRequest, BulkDeleteResponse
+#   from ai_ready_rag.schemas.admin import BulkDeleteRequest
+
+__all__ = [
+    # auth
+    "LoginRequest",
+    "LoginResponse",
+    "SetupRequest",
+    "UserBasicResponse",
+    # setup
+    "CompleteSetupRequest",
+    "CompleteSetupResponse",
+    "SetupStatusResponse",
+    # user
+    "TagAssignment",
+    "UserCreate",
+    "UserResponse",
+    "UserUpdate",
+    # tag
+    "TagCreate",
+    "TagResponse",
+    "TagUpdate",
+    # chat (excluding BulkDeleteRequest/BulkDeleteResponse)
+    "ConfidenceInfo",
+    "DeleteSessionResponse",
+    "MessageCreate",
+    "MessageListResponse",
+    "MessageResponse",
+    "SendMessageResponse",
+    "SessionCreate",
+    "SessionListItem",
+    "SessionListResponse",
+    "SessionResponse",
+    "SessionUpdate",
+    "SourceInfo",
+    # document (excluding BulkDeleteRequest/BulkDeleteResponse)
+    "BulkReprocessRequest",
+    "BulkReprocessResponse",
+    "CheckDuplicatesRequest",
+    "CheckDuplicatesResponse",
+    "DocumentListResponse",
+    "DocumentResponse",
+    "DuplicateErrorDetail",
+    "DuplicateInfo",
+    "ReprocessRequest",
+    "TagInfo",
+    "TagUpdateRequest",
+    # admin (excluding BulkDeleteRequest)
+    "AdvancedSettingsRequest",
+    "AdvancedSettingsResponse",
+    "ArchitectureInfoResponse",
+    "CacheClearResponse",
+    "CacheSeedRequest",
+    "CacheSeedResponse",
+    "CacheSettingsRequest",
+    "CacheSettingsResponse",
+    "CacheStatsResponse",
+    "CacheWarmRequest",
+    "CacheWarmResponse",
+    "ChangeEmbeddingRequest",
+    "ChangeEmbeddingResponse",
+    "ChangeModelRequest",
+    "ChangeModelResponse",
+    "ChatModelInfo",
+    "ClearKnowledgeBaseRequest",
+    "ClearKnowledgeBaseResponse",
+    "ComponentHealth",
+    "CuratedQACreate",
+    "CuratedQAListResponse",
+    "CuratedQAResponse",
+    "CuratedQAUpdate",
+    "DetailedHealthResponse",
+    "DocumentParsingInfo",
+    "EasyOCRStatus",
+    "EmbeddingsInfo",
+    "FileStats",
+    "InfrastructureStatus",
+    "KnowledgeBaseSummary",
+    "KnowledgeBaseStatsResponse",
+    "LLMSettingsRequest",
+    "LLMSettingsResponse",
+    "ManualWarmingRequest",
+    "ModelInfo",
+    "ModelLimits",
+    "ModelLimitsResponse",
+    "ModelsResponse",
+    "OCRStatus",
+    "ProcessingOptionsRequest",
+    "ProcessingOptionsResponse",
+    "ProcessingQueueStatus",
+    "RAGPipelineStatus",
+    "RecoverResponse",
+    "ReindexEstimate",
+    "ReindexFailureInfo",
+    "ReindexFailuresResponse",
+    "ReindexJobResponse",
+    "ResumeReindexRequest",
+    "RetrievalSettingsRequest",
+    "RetrievalSettingsResponse",
+    "SettingsAuditEntry",
+    "SettingsAuditResponse",
+    "StartReindexRequest",
+    "SynonymCreate",
+    "SynonymListResponse",
+    "SynonymResponse",
+    "SynonymUpdate",
+    "TesseractStatus",
+    "TopQueryItem",
+    "TopQueryResponse",
+    "WarmFileResponse",
+    "WarmingJobListResponse",
+    "WarmingJobResponse",
+    "WarmingQueueJobResponse",
+    "WarmingQueueListResponse",
+    "WarmRetryRequest",
+    # experimental
+    "GeneratePresentationRequest",
+    "PresentationResponse",
+    "SlideResponse",
+]
